feat: TBTI 관련 기능별로 Controller 분리
